Address Copilot PR #192 review nits
- mobile_controls.gd: drop unused SHOP_PATH_PREFIXES constant — never
  wired up after the polling switched to SceneManager.can_pop().
- mobile_controls.gd: replace `"is_pause_menu_open" in gs` GDScript
  membership check with a direct property access. GameState always
  exposes the field; the existence check was non-idiomatic and would
  fall through silently if the property name changed.
- web/src/utils/clipboard.ts: extract the HTTPS-fallback copyText
  helper out of ContentTab so the shared CellInspector (Layout tab)
  can use it too. The Layout tab's "Copy cell as JSON" button was
  still hitting navigator.clipboard.writeText directly and failing
  on plain HTTP. Same helper now also covers the gate-position
  click-to-copy in CellInspector.
