+++ web/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/character-creator/CharacterCreator.tsx","./src/character-creator/CharacterPreview.tsx","./src/character-creator/data/classData.ts","./src/character-creator/data/constants.ts","./src/character-creator/hooks/useCharacterState.ts","./src/character-creator/steps/AppearanceStep.tsx","./src/character-creator/steps/ClassStep.tsx","./src/character-creator/steps/ConfirmStep.tsx","./src/character-creator/steps/NameStep.tsx","./src/controls/ControlsDiagram.tsx","./src/elements/AreaWarp.tsx","./src/elements/BearTrap.tsx","./src/elements/Box.tsx","./src/elements/DropArmor.tsx","./src/elements/DropItem.tsx","./src/elements/DropMeseta.tsx","./src/elements/DropRare.tsx","./src/elements/DropWeapon.tsx","./src/elements/DropshipCrash.tsx","./src/elements/Fence.tsx","./src/elements/Fence4.tsx","./src/elements/Gate.tsx","./src/elements/InteractSwitch.tsx","./src/elements/Key.tsx","./src/elements/KeyGate.tsx","./src/elements/MessagePack.tsx","./src/elements/NeedleTrap.tsx","./src/elements/NpcKai.tsx","./src/elements/NpcSarisa.tsx","./src/elements/RareBox.tsx","./src/elements/RemoteSwitch.tsx","./src/elements/StartWarp.tsx","./src/elements/StepSwitch.tsx","./src/elements/Wall.tsx","./src/elements/Waypoint.tsx","./src/elements/index.ts","./src/elements/types.ts","./src/hooks/useMediaQuery.ts","./src/office-editor/OfficeCanvas.tsx","./src/office-editor/OfficeEditor.tsx","./src/office-editor/OfficeExportPanel.tsx","./src/office-editor/OfficeOverlay.tsx","./src/office-editor/types.ts","./src/pages/Landing.tsx","./src/photo-mode/PhotoMode.tsx","./src/quest-editor/CompanionPreview.tsx","./src/quest-editor/QuestEditor.tsx","./src/quest-editor/QuestHome.tsx","./src/quest-editor/constants.ts","./src/quest-editor/types.ts","./src/quest-editor/hooks/useQuestProject.ts","./src/quest-editor/hooks/useStageConfigs.ts","./src/quest-editor/preview/PreviewGrid.tsx","./src/quest-editor/preview/PreviewMinimap.tsx","./src/quest-editor/preview/StageScene.tsx","./src/quest-editor/shared/CellInspector.tsx","./src/quest-editor/shared/GridCanvas.tsx","./src/quest-editor/shared/StagePicker.tsx","./src/quest-editor/shared/grid-generation.ts","./src/quest-editor/tabs/ContentTab.tsx","./src/quest-editor/tabs/ExportTab.tsx","./src/quest-editor/tabs/LayoutTab.tsx","./src/quest-editor/tabs/MetadataTab.tsx","./src/quest-editor/tabs/PreviewTab.tsx","./src/quest-editor/utils/quest-io.ts","./src/retarget/RetargetTuner.tsx","./src/retarget/RetargetViewer.tsx","./src/retarget/retarget-utils.ts","./src/settings/SettingsMockup.tsx","./src/sfx-labeler/SfxLabeler.tsx","./src/shared/svg-minimap.ts","./src/stage-editor/FloorOverlay.tsx","./src/stage-editor/ObstacleOverlay.tsx","./src/stage-editor/ParticleOverlay.tsx","./src/stage-editor/PortalOverlay.tsx","./src/stage-editor/StageCanvas.tsx","./src/stage-editor/StageSelector.tsx","./src/stage-editor/TextureAnimator.tsx","./src/stage-editor/UnifiedStageEditor.tsx","./src/stage-editor/constants.ts","./src/stage-editor/types.ts","./src/stage-editor/useStageConfig.ts","./src/stage-editor/tabs/ExportTab.tsx","./src/stage-editor/tabs/FloorCollisionTab.tsx","./src/stage-editor/tabs/ObstacleTab.tsx","./src/stage-editor/tabs/PortalTab.tsx","./src/stage-editor/tabs/SceneTab.tsx","./src/stage-editor/tabs/SvgTab.tsx","./src/stage-editor/tabs/TextureTab.tsx","./src/start-menu/StartMenu.tsx","./src/storybook/BasicWeaponPreview.tsx","./src/storybook/EnemyGallery.tsx","./src/storybook/MenuDesign.tsx","./src/storybook/PlayerAnimationStorybook.tsx","./src/storybook/StorybookViewer.tsx","./src/storybook/WeaponGallery.tsx","./src/storybook/enemyData.ts","./src/storybook/weaponData.ts","./src/svg-check/SvgCheck.tsx","./src/title-screen/TitleScreen.tsx","./src/utils/ErrorBoundary.tsx","./src/utils/assets.ts"],"version":"5.7.3"}
+{"root":["./src/App.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/character-creator/CharacterCreator.tsx","./src/character-creator/CharacterPreview.tsx","./src/character-creator/data/classData.ts","./src/character-creator/data/constants.ts","./src/character-creator/hooks/useCharacterState.ts","./src/character-creator/steps/AppearanceStep.tsx","./src/character-creator/steps/ClassStep.tsx","./src/character-creator/steps/ConfirmStep.tsx","./src/character-creator/steps/NameStep.tsx","./src/controls/ControlsDiagram.tsx","./src/elements/AreaWarp.tsx","./src/elements/BearTrap.tsx","./src/elements/Box.tsx","./src/elements/DropArmor.tsx","./src/elements/DropItem.tsx","./src/elements/DropMeseta.tsx","./src/elements/DropRare.tsx","./src/elements/DropWeapon.tsx","./src/elements/DropshipCrash.tsx","./src/elements/Fence.tsx","./src/elements/Fence4.tsx","./src/elements/Gate.tsx","./src/elements/InteractSwitch.tsx","./src/elements/Key.tsx","./src/elements/KeyGate.tsx","./src/elements/MessagePack.tsx","./src/elements/NeedleTrap.tsx","./src/elements/NpcKai.tsx","./src/elements/NpcSarisa.tsx","./src/elements/RareBox.tsx","./src/elements/RemoteSwitch.tsx","./src/elements/StartWarp.tsx","./src/elements/StepSwitch.tsx","./src/elements/Wall.tsx","./src/elements/Waypoint.tsx","./src/elements/index.ts","./src/elements/types.ts","./src/hooks/useMediaQuery.ts","./src/office-editor/OfficeCanvas.tsx","./src/office-editor/OfficeEditor.tsx","./src/office-editor/OfficeExportPanel.tsx","./src/office-editor/OfficeOverlay.tsx","./src/office-editor/types.ts","./src/pages/Landing.tsx","./src/photo-mode/PhotoMode.tsx","./src/quest-editor/CompanionPreview.tsx","./src/quest-editor/QuestEditor.tsx","./src/quest-editor/QuestHome.tsx","./src/quest-editor/constants.ts","./src/quest-editor/types.ts","./src/quest-editor/hooks/useQuestProject.ts","./src/quest-editor/hooks/useStageConfigs.ts","./src/quest-editor/preview/PreviewGrid.tsx","./src/quest-editor/preview/PreviewMinimap.tsx","./src/quest-editor/preview/StageScene.tsx","./src/quest-editor/shared/CellInspector.tsx","./src/quest-editor/shared/GridCanvas.tsx","./src/quest-editor/shared/StagePicker.tsx","./src/quest-editor/shared/grid-generation.ts","./src/quest-editor/tabs/ContentTab.tsx","./src/quest-editor/tabs/ExportTab.tsx","./src/quest-editor/tabs/LayoutTab.tsx","./src/quest-editor/tabs/MetadataTab.tsx","./src/quest-editor/tabs/PreviewTab.tsx","./src/quest-editor/utils/quest-io.ts","./src/retarget/RetargetTuner.tsx","./src/retarget/RetargetViewer.tsx","./src/retarget/retarget-utils.ts","./src/settings/SettingsMockup.tsx","./src/sfx-labeler/SfxLabeler.tsx","./src/shared/svg-minimap.ts","./src/stage-editor/FloorOverlay.tsx","./src/stage-editor/ObstacleOverlay.tsx","./src/stage-editor/ParticleOverlay.tsx","./src/stage-editor/PortalOverlay.tsx","./src/stage-editor/StageCanvas.tsx","./src/stage-editor/StageSelector.tsx","./src/stage-editor/TextureAnimator.tsx","./src/stage-editor/UnifiedStageEditor.tsx","./src/stage-editor/constants.ts","./src/stage-editor/types.ts","./src/stage-editor/useStageConfig.ts","./src/stage-editor/tabs/ExportTab.tsx","./src/stage-editor/tabs/FloorCollisionTab.tsx","./src/stage-editor/tabs/ObstacleTab.tsx","./src/stage-editor/tabs/PortalTab.tsx","./src/stage-editor/tabs/SceneTab.tsx","./src/stage-editor/tabs/SvgTab.tsx","./src/stage-editor/tabs/TextureTab.tsx","./src/start-menu/StartMenu.tsx","./src/storybook/BasicWeaponPreview.tsx","./src/storybook/EnemyGallery.tsx","./src/storybook/MenuDesign.tsx","./src/storybook/PlayerAnimationStorybook.tsx","./src/storybook/StorybookViewer.tsx","./src/storybook/WeaponGallery.tsx","./src/storybook/enemyData.ts","./src/storybook/weaponData.ts","./src/svg-check/SvgCheck.tsx","./src/title-screen/TitleScreen.tsx","./src/utils/ErrorBoundary.tsx","./src/utils/assets.ts","./src/utils/clipboard.ts"],"version":"5.7.3"}
